fix: 修改统计提现同期群数据

+++ co-driver-statistics/co-driver-statistics-services/co-driver-statistics-service-dwd/src/main/java/com/leo/ad/codriver/dwd/entity/DwdUserWithdrawRecord.java
@@ -39,8 +39,8 @@ public class DwdUserWithdrawRecord implements BaseEntity {
     private Long registerCohortDay;
     private Date createTime;
     private Integer updateDates;
-    private Long registerWithdrawDay;
-    private Long registerWithdrawCohortDay;
+    private Integer registerWithdrawDay;
+    private Integer registerWithdrawCohortDay;
     private Date updateTime;
     @TableField(exist = false)
     private String userSource;
